Add support for --version/-V option
This change adds support for the --version/-V option(s) to the program.
When used, the program's version (including a git revision, if
available) will be printed to stdout. In order to determine the git
revision at which the program was built, we rely on the newly added grev
build dependency.