Detect fullscreen apps
There are cases when a notification is non-desired while watching a
video, for example.
With this change we make an attempt at detecting such usage by checking
if the currently active window is running in fullscreen mode. If that is
the case, we set back the goggle time just as we do when exceeding the
user configurable idle reset duration.