Introduce run_once() function
With this change we refactor the code further, moving the sleep
functionality from the main loop into the check_once() function and
renaming the same to run_once(). Co-locating the sleep with the rest of
the logic will be beneficial in the future, once we switch from
remembering the time stamp since when the user goggled to using a
duration.