Switch from working with an Instant to using a Duration
So far we simply remembered the Instant at which the user started
goggling and used the difference to the current time to decided whether
a notification is due. However, working with a time stamp is rather
limiting, because it does not allow us to implement any form of pause
very naturally.
To allow for such functionality, this change switches over to using a
Duration instead of an Instant.