Introduce Daemon struct
We may need more state to be present on each tick, but explicitly
passing it back and forth between run_once() and the main loop is
cumbersome.
With this change we introduce a Daemon struct that contains the relevant
state and that now homes the run() and run_once() functionality in the
form of methods.