Pause goggle duration when window is in fullscreen mode
It's arguably not great to reset the current goggle duration just
because an application is detected to be running in fullscreen mode:
this can easily happen when a user is briefly playing a game or watching
a video.
The more appropriate behavior is to pause the goggle duration for the
time being.