Pause goggle time when no window is active
Our current logic assumes that there is always an active window.
However, as it turns out the window 0x0 (which we see being reported),
is not actually valid and is being used to indicate that no window is
active.
With this change we take into account whether or not a window is active.
If no window is active we may end up pausing goggle time accumulation
temporarily.