Only send notification if not currently idle
There conceivably is a race condition between the daemon sending a
notification to remind the user to take a break while the user did just
that. In such a case the notification is emitted needlessly and may be
perceived as a nuisance once the user returns.
To prevent such cases from happening, this change make sure that the
user has not been idle for a while before sending said notification.